Analysis of the Relative Quantifications

Analysis of relative quantifications, including:

Annotations

Summary files in different format (xls, txt) and shapes (long, wide)

Numerous summary plots

Enrichment analysis using Gprofiler

PCA of quantifications

Clustering analysis

Basic imputation of missing values

artmsAnalysisQuantifications(
  log2fc_file,
  modelqc_file,
  species,
  output_dir = "analysis_quant",
  outliers = c("keep", "iqr", "std"),
  enrich = TRUE,
  l2fc_thres = 1,
  choosePvalue = c("adjpvalue", "pvalue"),
  isBackground = "nobackground",
  isPtm = "global",
  mnbr = 2,
  isFluomics = FALSE,
  pathogen = "nopathogen",
  plotPvaluesLog2fcDist = TRUE,
  plotAbundanceStats = TRUE,
  plotReproAbundance = TRUE,
  plotCorrConditions = TRUE,
  plotCorrQuant = TRUE,
  plotPCAabundance = TRUE,
  plotFinalDistributions = TRUE,
  plotPropImputation = TRUE,
  plotHeatmapsChanges = TRUE,
  plotTotalQuant = TRUE,
  plotClusteringAnalysis = TRUE,
  data_object = FALSE,
  verbose = TRUE
)

Arguments

log2fc_file
(char) MSstats results file location

modelqc_file
(char) MSstats modelqc file location

species
(char) Select one species. Species currently supported for a full analysis (including enrichment analysis):

HUMAN

MOUSE

To find out species supported only for annotation check ?artmsIsSpeciesSupported()

output_dir
(char) Name for the folder to output the results from the function. Default is current directory (recommended to provide a new folder name).

outliers
(char) It allows to keep or remove outliers. Options:

keep (default): it keeps outliers 'keep', 'iqr', 'std'

iqr (recommended): remove outliers +/- 6 x Interquartile Range (IQR)

std : 6 x standard deviation

enrich
(logical) Performed enrichment analysis using GprofileR? Only available for species HUMAN and MOUSE. TRUE (default if "human" or "mouse" are the species) or FALSE

l2fc_thres
(int) log2fc cutoff for enrichment analysis (default, l2fc_thres = 1.5)

choosePvalue
(char) specify whether pvalue or adjpvalue should use for the analysis. The default option is adjpvalue (multiple testing correction). But if the number of biological replicates for a given experiment is too low (for example n = 2), then choosePvalue = pvalue is recommended.

isBackground
(char) background of gene names for enrichment analysis. nobackground (default) will use the total number of genes detected. Alternatively provided the file path name to the background gene list.

isPtm
(char) Is a ptm-site quantification?

global (default),

ptmsites (for site specific analysis),

ptmph (Jeff Johnson script output evidence file)

mnbr
(int) minimal number of biological replicates for imputation and filtering. Default: mnbr = 2 (Proteins must be found in one of the conditions in at least 2 of the biological replicates)

isFluomics
(logical) Does this data belong to the FluOMICs project? TRUE or FALSE (default)

pathogen
(char) Is there a pathogen in the dataset as well? if it does not, then use pathogen = nopathogen (default). Pathogens available: tb (Tuberculosis), lpn (Legionella)

plotPvaluesLog2fcDist
(logical) If TRUE (default) plots pvalues and log2fc distributions

plotAbundanceStats
(logical) If TRUE (default) plots stats graphs about abundance values

plotReproAbundance
(logical) If TRUE plots reproducibility based on normalized abundance values

plotCorrConditions
(logical) If TRUE plots correlation between the different conditions

plotCorrQuant
(logical) if TRUE plots correlation between the available quantifications (comparisons)

plotPCAabundance
(logical) if TRUE performs PCA analysis of conditions using normalized abundance values

plotFinalDistributions
(logical) if TRUE plots distribution of both log2fc and pvalues

plotPropImputation
(logical) if TRUE plots proportion of overall imputation

plotHeatmapsChanges
(logical) if TRUE plots heatmaps of quantified changes (both all and significant only)

plotTotalQuant
(logical) if TRUE plots barplot of total number of quantifications per comparison

plotClusteringAnalysis
(logical) if TRUE performs clustering analysis between quantified comparisons (more than 1 comparison required)

data_object
(logical) flag to indicate whether the required files are data objects. Default is FALSE

verbose
(logical) TRUE (default) shows function messages

Value

(data.frame) summary of quantifications, including annotations, enrichments, etc

Examples

# Testing that the files cannot be empty
artmsAnalysisQuantifications(log2fc_file = NULL,
                              modelqc_file = NULL,
                              species = NULL,
                              output_dir = NULL)

#> ---------------------------------------------

#> artMS: ANALYSIS OF QUANTIFICATIONS

#> ---------------------------------------------

#> [1] "The evidence_file, modelqc_file, species and output_dir arguments cannot be NULL"